refactor: add constants folder
