+++ .eslintrc
@@ -3,7 +3,7 @@
     "airbnb-base"
   ],
   "rules": {
-    "max-len": [2, 80, 4, {"ignoreUrls": true}],
+    "max-len": [2, 120, 4, {"ignoreUrls": true}],
     "no-param-reassign": 0,
     "indent": "off"
   },
