+++ src/js/constants/keyboardConstants.js
@@ -1,0 +1,35 @@
+export default {
+    shiftKeyboardLayout: {
+        RU: ['Ё', '!', '"', '№', ';', '%', ':', '?', '*', '(',
+            ')', '_', '+', '{', '}', ':', '"', '<', '>', '?'],
+        ENG: ['~', '!', '@', '#', '$', '%', '^', '&', '*', '(',
+            ')', '_', '+', '{', '}', ':', '"', '<', '>', '?'],
+        DEFAULT: ['`', '1', '2', '3', '4', '5', '6', '7', '8',
+            '9', '0', '-', '=', '[', ']', ';', '', ',', '.', '/'],
+    },
+
+    KEY_CODES: ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8',
+        'Digit9', 'Digit0', 'Minus', 'Equal', 'Backspace', 'Tab', 'KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT',
+        'KeyY', 'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Delete', 'CapsLock', 'KeyA',
+        'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Enter', 'ShiftLeft',
+        'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'Slash', 'ShiftRight', 'ArrowUp',
+        'ControlLeft', 'MetaLeft', 'AltLeft', 'Space', 'AltRight', 'ControlRight', 'ArrowLeft',
+        'ArrowDown', 'ArrowRight'],
+
+    CONTROL_BUTTONS: ['Control', 'CapsLock', 'Shift', 'Win', 'Del', 'Alt', 'Backspace', 'Tab', 'Enter', 'Space'],
+
+    languageKeyboard: {
+        RU: ['ё', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', 'Backspace',
+            'Tab', 'й', 'ц', 'у', 'к', 'е', 'н', 'г', 'ш', 'щ', 'з', 'х', 'ъ', 'Del',
+            'CapsLock', 'ф', 'ы', 'в', 'а', 'п', 'р', 'о', 'л', 'д', 'ж', 'э', 'Enter',
+            'Shift', 'я', 'ч', 'с', 'м', 'и', 'т', 'ь', 'б', 'ю', '.', 'Shift', '↑',
+            'Control', 'Win', 'Alt', 'Space', 'Alt', 'Control', '←', '↓', '→',
+        ],
+        ENG: ['`', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', 'Backspace',
+            'Tab', 'q', 'w', 'e', 'r', 't', 'y', 'u', 'i', 'o', 'p', '[', ']', 'Del',
+            'CapsLock', 'a', 's', 'd', 'f', 'g', 'h', 'j', 'k', 'l', ';', "'", 'Enter',
+            'Shift', 'z', 'x', 'c', 'v', 'b', 'n', 'm', ',', '.', '/', 'Shift', '↑',
+            'Control', 'Win', 'Alt', 'Space', 'Alt', 'Control', '←', '↓', '→',
+        ],
+    },
+};
