+++ src/js/constants/layoutConstants.js
@@ -1,0 +1,5 @@
+export default {
+    KEYBOARD: document.createElement('div'),
+    INFO_ABOUT_US: document.createElement('div'),
+    TEXT_AREA: document.createElement('textarea'),
+};
